update console colors
Note that this doesn't seem to get picked up by Eclipse on first run.
The colors will be set, but they will not be applied, forcing you to
change each one to something else, click "Apply", then change them back
again and clicking "Apply" again. Guessing this is a bug in Eclipse.